+++ src/Config/ConfigFormGUI.php
@@ -46,12 +46,25 @@ class ConfigFormGUI extends ActiveRecordConfigFormGUI {
                 self::PROPERTY_CLASS => ilCheckboxInputGUI::class,
                 "setTitle"           => self::plugin()->translate(Config::KEY_SHOW_MATRIX_ACTIONS, self::LANG_MODULE, [self::dic()->language()->txt("trac_matrix")]),
                 ],
-            Config::KEY_SHOW_ONLY_APPEARABLE_ORG_UNITS_IN_FILTER => [
-                self::PROPERTY_CLASS => ilCheckboxInputGUI::class
-            ],
             Config::KEY_ENABLE_COURSES_VIEW => [
                 self::PROPERTY_CLASS => ilCheckboxInputGUI::class,
                 "setTitle"           => self::plugin()->translate("enable_view", self::LANG_MODULE, [self::dic()->language()->txt("courses")])
+            ],
+            Config::KEY_ENABLE_REPORTING_VIEW => [
+                self::PROPERTY_CLASS    => ilCheckboxInputGUI::class,
+                self::PROPERTY_SUBITEMS => [
+                    Config::KEY_ENABLE_REPORTING_VIEW_PER_OBJECT => [
+                        self::PROPERTY_CLASS    => ilCheckboxInputGUI::class,
+                        self::PROPERTY_SUBITEMS => [
+                            Config::KEY_ENABLE_REPORTING_VIEW_PER_OBJECT_NEW_OBJECTS => [
+                                self::PROPERTY_CLASS => ilCheckboxInputGUI::class,
+                                "setTitle"           => self::plugin()->translate("reporting_per_object_new_objects", self::LANG_MODULE)
+                            ]
+                        ],
+                        "setTitle"              => self::plugin()->translate("reporting_per_object", self::LANG_MODULE)
+                    ]
+                ],
+                "setTitle"              => self::plugin()->translate("reporting", self::LANG_MODULE)
             ],
             Config::KEY_ENABLE_USERS_VIEW   => [
                 self::PROPERTY_CLASS => ilCheckboxInputGUI::class,
@@ -75,6 +88,9 @@ class ConfigFormGUI extends ActiveRecordConfigFormGUI {
             self::PROPERTY_CLASS => ilCheckboxInputGUI::class,
             "setTitle" => self::dic()->language()->txt("trac_learning_progress") . " " . self::dic()->language()->txt("notes_comments")
         ],
+            Config::KEY_SYNC_POSITION_PERMISSIONS_WITH_CHILDREN => [
+                self::PROPERTY_CLASS => ilCheckboxInputGUI::class
+            ]
 		];
 	}
 }