+++ src/Report/Matrix/MatrixTableGUI.php
@@ -343,5 +343,7 @@ class MatrixTableGUI extends AbstractReport2TableGUI
     protected function extendsActionsMenu(ilAdvancedSelectionListGUI $actions, array $row)/*: void*/
     {
         self::dic()->ctrl()->setParameter($this->parent_obj, Reports::GET_PARAM_USR_ID, $row["usr_id"]);
+
+        $actions->setId($row["usr_id"]);
     }
 }